Move common code out
