+++ clang/lib/ScalableStaticAnalysisFramework/CMakeLists.txt
@@ -1,4 +1,4 @@
 add_subdirectory(Analyses)
 add_subdirectory(Core)
 add_subdirectory(Frontend)
-
+add_subdirectory(Tool)
