+++ clang/lib/ScalableStaticAnalysisFramework/Tool/CMakeLists.txt
@@ -1,0 +1,11 @@
+set(LLVM_LINK_COMPONENTS
+  Support
+  )
+
+add_clang_library(clangScalableStaticAnalysisFrameworkTool
+  Utils.cpp
+
+  LINK_LIBS
+  clangBasic
+  clangScalableStaticAnalysisFrameworkCore
+  )
